+++ noodles-sam/src/reader/record/quality_scores.rs
@@ -5,6 +5,8 @@ use crate::record::{quality_scores, QualityScores};
 /// An error when raw SAM record quality scores fail to parse.
 #[derive(Clone, Debug, Eq, PartialEq)]
 pub enum ParseError {
+    /// The input is empty.
+    Empty,
     /// The input is invalid.
     Invalid(quality_scores::ParseError),
 }
@@ -12,6 +14,7 @@ pub enum ParseError {
 impl error::Error for ParseError {
     fn source(&self) -> Option<&(dyn error::Error + 'static)> {
         match self {
+            Self::Empty => None,
             Self::Invalid(e) => Some(e),
         }
     }
@@ -20,6 +23,7 @@ impl error::Error for ParseError {
 impl fmt::Display for ParseError {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         match self {
+            Self::Empty => write!(f, "empty input"),
             Self::Invalid(_) => write!(f, "invalid input"),
         }
     }
@@ -30,6 +34,10 @@ pub(crate) fn parse_quality_scores(
     quality_scores: &mut QualityScores,
 ) -> Result<(), ParseError> {
     const OFFSET: u8 = b'!';
+
+    if src.is_empty() {
+        return Err(ParseError::Empty);
+    }
 
     let raw_quality_scores = Vec::from(mem::take(quality_scores));
 
@@ -50,13 +58,15 @@ mod tests {
         let mut quality_scores = QualityScores::default();
 
         quality_scores.clear();
-        parse_quality_scores(b"", &mut quality_scores)?;
-        assert!(quality_scores.is_empty());
-
-        quality_scores.clear();
         parse_quality_scores(b"NDLS", &mut quality_scores)?;
         let expected = QualityScores::try_from(vec![45, 35, 43, 50])?;
         assert_eq!(quality_scores, expected);
+
+        quality_scores.clear();
+        assert_eq!(
+            parse_quality_scores(b"", &mut quality_scores),
+            Err(ParseError::Empty)
+        );
 
         quality_scores.clear();
         assert!(matches!(